cut into components
